# api/oas-doc.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Accounts Service
  description: YourYummy account manager service 
paths:
...
  /api/v1/refreshToken:
    x-router-controller: credentialManager
    post:
      x-swagger-ui: false
      description: Refreshes the access token
      tags: ["User Access"]
      operationId: refreshToken
      requestBody:
        description: Token to refresh
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        201: 
          description: Token refreshed
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          

  # ACCOUNT MANAGEMENT
...
components:
  schemas:
...
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
  securitySchemes:
    apikey:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-acl-config: { $ref: ./acl-config.yaml }